Extract `AndroidStudioProvisioningPlugin` from `PerformanceTestPlugin`
